refactor: update import paths and dependencies; bump package versions
