+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@ repos:
       - id: detect-private-key
       - id: check-docstring-first
   - repo: https://github.com/pycqa/isort
-    rev: 5.13.2
+    rev: 6.0.0
     hooks:
       - id: isort
         name: isort (python)
@@ -20,12 +20,12 @@ repos:
     hooks:
       - id: pyupgrade
   - repo: https://github.com/psf/black
-    rev: 24.10.0
+    rev: 25.1.0
     hooks:
       - id: black
         exclude: alembic/versions/
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v1.14.1'
+    rev: 'v1.15.0'
     hooks:
       - id: mypy
         exclude: ^alembic/
@@ -36,11 +36,11 @@ repos:
           - --install-types
           - --non-interactive
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.9.2
+    rev: v0.9.4
     hooks:
       - id: ruff
         exclude: ^alembic/
   - repo: https://github.com/astral-sh/uv-pre-commit
-    rev: 0.5.20
+    rev: 0.5.29
     hooks:
       - id: uv-lock
